Move nested classes
