+++ src/main/kotlin/com/github/hubbards/xpath/StringBuilder.kt
@@ -1,0 +1,13 @@
+package com.github.hubbards.xpath
+
+internal fun StringBuilder.parenthesize(string: String) {
+  append('(')
+  append(string)
+  append(')')
+}
+
+internal fun StringBuilder.brackets(string: String) {
+  append('[')
+  append(string)
+  append(']')
+}
