+++ src/test/java/com/github/hubbards/xpath/ExpressionJavaTest.java
@@ -6,22 +6,22 @@ import org.junit.Test;
 public class ExpressionJavaTest {
     @Test
     public void literalNumberSyntax() {
-        Expression expression = new Expression.LiteralNumber(3);
+        Expression expression = new LiteralNumber(3);
         Assert.assertEquals("3", expression.getUnabbreviated());
         Assert.assertEquals("3", expression.getAbbreviated());
     }
 
     @Test
     public void literalStringSyntax() {
-        Expression expression = new Expression.LiteralString("dog");
+        Expression expression = new LiteralString("dog");
         Assert.assertEquals("'dog'", expression.getUnabbreviated());
         Assert.assertEquals("'dog'", expression.getAbbreviated());
     }
 
     @Test
     public void functionCallSyntax() {
-        Expression expression = Expression.FunctionCall.Factory.localName(
-                Expression.Path.Factory.absolute(
+        Expression expression = FunctionCall.Factory.localName(
+                Path.Factory.absolute(
                         Step.Factory.step(Axis.DESCENDANT, new NodeTest.Name("img")),
                         Step.Factory.step(Axis.PARENT, NodeTest.Node.INSTANCE)
                 )
@@ -32,10 +32,10 @@ public class ExpressionJavaTest {
 
     @Test
     public void binaryExpressionSyntax() {
-        Expression expression = new Expression.BinaryExpression(
+        Expression expression = new BinaryExpression(
                 Operator.GREATER_THAN,
-                new Expression.LiteralNumber(3),
-                Expression.FunctionCall.Factory.count(Expression.Path.Factory.absolute(
+                new LiteralNumber(3),
+                FunctionCall.Factory.count(Path.Factory.absolute(
                         Step.Factory.step(Axis.DESCENDANT_OR_SELF, NodeTest.Node.INSTANCE),
                         Step.Factory.step(Axis.CHILD, new NodeTest.Name("ol")),
                         Step.Factory.step(Axis.CHILD, new NodeTest.Name("li"))
